Squashed commit of the following:
commit 6864011a0a433e362b168ce5ae5b7d36f532b80a
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Sun Nov 26 16:52:48 2023 +0100

    fixes

commit 36ede6b96d0028c730338e38b6242d8a222c641a
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Mon Nov 13 23:16:01 2023 +0100

    tests

commit 9db48223dcf108dc1da11b38494f2f228f34a04c
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Mon Nov 13 22:58:45 2023 +0100

    changes 2

commit e15bca415c2adda2cc39a6f26bb5661c96a4ff20
Author: Roman Paukner <romanpauk@gmail.com>
Date:   Mon Nov 13 22:37:36 2023 +0100

    changes

+++ examples/message_processing.cpp
@@ -18,7 +18,7 @@ struct MessageWrapper {
         messages_ = std::forward<T>(message);
     }
 
-    int id() const { return messages_.index(); }
+    size_t id() const { return messages_.index(); }
     const MessageA& GetA() const { return std::get<MessageA>(messages_); }
     const MessageB& GetB() const { return std::get<MessageB>(messages_); }
 
@@ -47,11 +47,11 @@ struct ProcessorB : IProcessor {
 int main() {
     using namespace dingo;
 
-    // Define traits type with a single index using int as a key,
+    // Define traits type with a single index using size_t as a key,
     // backed by a std::array of size 10
     struct container_traits : static_container_traits<void> {
         using index_definition_type =
-            std::tuple<std::tuple<int, index_type::array<10>>>;
+            std::tuple<std::tuple<size_t, index_type::array<10>>>;
     };
 
     container<container_traits> container;
@@ -59,10 +59,10 @@ int main() {
     // Register processors into the container, indexed by the type they process
     container.register_indexed_type<scope<shared>,
                                     storage<std::shared_ptr<ProcessorA>>,
-                                    interface<IProcessor>>(1);
+                                    interface<IProcessor>>(size_t(1));
     container.register_indexed_type<scope<unique>,
                                     storage<std::shared_ptr<ProcessorB>>,
-                                    interface<IProcessor>>(2);
+                                    interface<IProcessor>>(size_t(2));
 
     // Register repositories used by the processors
     container.register_type<scope<shared>, storage<RepositoryA>>();
